Changed cli to typescript
